Updaate tests and docs
