+++ tests/test_project_structure.py
@@ -142,7 +142,9 @@ class TestGoogleProviderProjectStructure(unittest.TestCase):
         ('cloud', 'mysql_to_gcs'),
         ('cloud', 'mssql_to_gcs'),
         ('cloud', 'bigquery_to_gcs'),
-        ('cloud', 'local_to_gcs')
+        ('cloud', 'local_to_gcs'),
+        ('cloud', 'sheets_to_gcs'),
+        ('suite', 'gcs_to_sheets'),
     }
 
     MISSING_DOC_GUIDES = {
